move `CodableFeedStore` to its own file in production
